requirements.txt
fastapi>=0.110.0
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
requests
openai
python-dotenv
pytest
httpx
loguru
anyio>=4.5.0
starlette>=0.37.0
openai-agents
aiohttp
ijson
voyageai
weaviate-client